movies_admin/requirements.txt
asgiref==3.7.2
attrs==23.2.0
black==23.12.1
cfgv==3.4.0
click==8.1.7
distlib==0.3.8
Django==4.2.5
django-debug-toolbar==4.2.0
django-split-settings==1.2.0
filelock==3.13.1
flake8==7.0.0
flake8-bugbear==24.1.17
identify==2.5.33
isort==5.13.2
mccabe==0.7.0
mypy-extensions==1.0.0
nodeenv==1.8.0
packaging==23.2
pathspec==0.12.1
pep8-naming==0.13.3
platformdirs==4.2.0
pre-commit==3.6.0
psycopg2==2.9.9
pycodestyle==2.11.1
pyflakes==3.2.0
python-dotenv==1.0.1
PyYAML==6.0.1
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.9.0
virtualenv==20.25.0
